[CH3] refactor : make AddStatement class to anonymous inner class
